feat: create rope.py, attention.py
