+++ models/llama3/main.py
@@ -10,11 +10,11 @@ def main():
   device = set_device()
   set_seed(device)
 
-  model = Llama.from_pretrained(max_batch_size=2, model_desc='8B', version=1).to(device)
+  model = Llama.from_pretrained(max_batch_size=1, model_desc='3B', version=2).to(device)
 
   prompts = [
     "Simply put, the theory of relativity states that",
-    "The phenomenon of global warming refers to the",
+    # "The phenomenon of global warming refers to the",
   ]
 
   out = model.text_completion(prompts, max_gen_len=64, temperature=0.9, echo=True)
